+++ lib/l10n/app_localizations_en.dart
@@ -294,6 +294,23 @@ class AppLocalizationsEn extends AppLocalizations {
   String get tabTechnical => 'Technical';
 
   @override
+  String tabValueCount(num count) {
+    final intl.NumberFormat countNumberFormat = intl.NumberFormat.compact(
+      locale: localeName,
+    );
+    final String countString = countNumberFormat.format(count);
+
+    String _temp0 = intl.Intl.pluralLogic(
+      count,
+      locale: localeName,
+      other: '$countString values',
+      one: '1 value',
+      zero: 'no values',
+    );
+    return '$_temp0';
+  }
+
+  @override
   String get tabWatchlist => 'Watchlist';
 
   @override